% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-pt.R
\name{hyph.support.pt}
\alias{hyph.support.pt}
\title{Language support for Portuguese}
\usage{
hyph.support.pt()
}
\description{
This function adds support for Portuguese to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.pt}}).
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
